+++ Packages/Visuals/Parallax/Runtime/BFParallaxAxisLock.cs
@@ -1,0 +1,9 @@
+namespace BFTools.Visuals.Parallax
+{
+    public enum BFParallaxAxisLock
+    {
+        None,
+        PositiveOnly,
+        NegativeOnly
+    }
+}