+++ Packages/Visuals/Parallax/Runtime/BFParallaxLayerConfig.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace BFTools.Visuals.Parallax
+{
+    public abstract class BFParallaxLayerConfig : ScriptableObject
+    {
+        public abstract IBFParallaxLayer CreateLayer();
+    }
+}