+++ Packages/Visuals/Parallax/Runtime/IBFParallaxLayer.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace BFTools.Visuals.Parallax
+{
+    public interface IBFParallaxLayer
+    {
+        void Init(Transform parent, int sortingOrder);
+        void Tick(Vector2 cameraDisplacement, float dt);
+        void Cleanup();
+    }
+}